requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
Flask-Script==2.0.6
flask-marshmallow==0.14.0
marshmallow==3.8.0
flake8>=3.7
peewee>=3.12
PyMySQL>=0.9.3
pytz>=2019.3
tzlocal>=2.0
PyJWT>=1.7.1
Werkzeug>=1.0.1
